+++ tasks/git/src/main/java/com/walmartlabs/concord/plugins/git/GitTask.java
@@ -1,4 +1,5 @@
 package com.walmartlabs.concord.plugins.git;
+
 /*-
  * *****
  * Concord